+++ codeanalyzer/options/__init__.py
@@ -1,3 +1,15 @@
-from .options import AnalysisOptions, EmitTarget, OutputFormat, ShardStrategy
+from .options import (
+    AnalysisOptions,
+    CallGraphBackend,
+    EmitTarget,
+    OutputFormat,
+    ShardStrategy,
+)
 
-__all__ = ["AnalysisOptions", "EmitTarget", "OutputFormat", "ShardStrategy"]
+__all__ = [
+    "AnalysisOptions",
+    "CallGraphBackend",
+    "EmitTarget",
+    "OutputFormat",
+    "ShardStrategy",
+]